+++ fem/bilininteg_ext.cpp
@@ -1,0 +1,650 @@
+// Copyright (c) 2010, Lawrence Livermore National Security, LLC. Produced at
+// the Lawrence Livermore National Laboratory. LLNL-CODE-443211. All Rights
+// reserved. See file COPYRIGHT for details.
+//
+// This file is part of the MFEM library. For more information and source code
+// availability see http://mfem.org.
+//
+// MFEM is free software; you can redistribute it and/or modify it under the
+// terms of the GNU Lesser General Public License (as published by the Free
+// Software Foundation) version 2.1 dated February 1999.
+
+#include "../config/config.hpp"
+
+#include "../general/okina.hpp"
+
+#include "fem.hpp"
+#include <map>
+#include <cmath>
+#include <algorithm>
+#include "bilininteg.hpp"
+#include "bilininteg_ext.hpp"
+#include "kernels/mass.hpp"
+#include "kernels/diffusion.hpp"
+#include "../linalg/device.hpp"
+#include "../linalg/kernels/vector.hpp"
+#include "./kernels/geom.hpp"
+
+using namespace std;
+
+namespace mfem
+{
+
+// *****************************************************************************
+static const IntegrationRule &DefaultGetRule(const FiniteElement &trial_fe,
+                                             const FiniteElement &test_fe)
+{
+   int order;
+   if (trial_fe.Space() == FunctionSpace::Pk)
+   {
+      order = trial_fe.GetOrder() + test_fe.GetOrder() - 2;
+   }
+   else
+   {
+      // order = 2*el.GetOrder() - 2;  // <-- this seems to work fine too
+      order = trial_fe.GetOrder() + test_fe.GetOrder() + trial_fe.GetDim() - 1;
+   }
+   if (trial_fe.Space() == FunctionSpace::rQk)
+   {
+      return RefinedIntRules.Get(trial_fe.GetGeomType(), order);
+   }
+   return IntRules.Get(trial_fe.GetGeomType(), order);
+}
+
+// *****************************************************************************
+// * PADiffusionIntegrator
+// *****************************************************************************
+void PADiffusionIntegrator::Assemble(const FiniteElementSpace &fes)
+{
+   const Mesh *mesh = fes.GetMesh();
+   const IntegrationRule *rule = IntRule;
+   const FiniteElement &el = *fes.GetFE(0);
+   const IntegrationRule *ir = rule?rule:&DefaultGetRule(el,el);
+   const int dims = el.GetDim();
+   const int symmDims = (dims * (dims + 1)) / 2; // 1x1: 1, 2x2: 3, 3x3: 6
+   const int nq = ir->GetNPoints();
+   dim = mesh->Dimension();
+   ne = fes.GetNE();
+   dofs1D = el.GetOrder() + 1;
+   quad1D = IntRules.Get(Geometry::SEGMENT, ir->GetOrder()).GetNPoints();
+   const GeometryExtension *geo = GeometryExtension::Get(fes,*ir);
+   maps = DofToQuad::Get(fes, fes, *ir);
+   vec.SetSize(symmDims * nq * ne);
+   const double coeff = static_cast<ConstantCoefficient*>(Q)->constant;
+   kernels::fem::DiffusionAssemble(dim, quad1D, ne,
+                                   maps->W,
+                                   geo->J,
+                                   coeff,
+                                   vec);
+   delete geo;
+}
+
+// *****************************************************************************
+void PADiffusionIntegrator::MultAdd(Vector &x, Vector &y)
+{
+   kernels::fem::DiffusionMultAssembled(dim, dofs1D, quad1D, ne,
+                                        maps->B,
+                                        maps->G,
+                                        maps->Bt,
+                                        maps->Gt,
+                                        vec, x, y);
+}
+
+
+// *****************************************************************************
+// * PAMassIntegrator
+// *****************************************************************************
+void PAMassIntegrator::Assemble(const FiniteElementSpace &fes)
+{
+   const Mesh *mesh = fes.GetMesh();
+   const IntegrationRule *rule = IntRule;
+   const FiniteElement &el = *fes.GetFE(0);
+   const IntegrationRule *ir = rule?rule:&DefaultGetRule(el,el);
+   dim = mesh->Dimension();
+   ne = fes.GetMesh()->GetNE();
+   nq = ir->GetNPoints();
+   dofs1D = el.GetOrder() + 1;
+   quad1D = IntRules.Get(Geometry::SEGMENT, ir->GetOrder()).GetNPoints();
+   const GeometryExtension *geo = GeometryExtension::Get(fes,*ir);
+   maps = DofToQuad::Get(fes, fes, *ir);
+   vec.SetSize(ne*nq);
+   ConstantCoefficient *const_coeff = dynamic_cast<ConstantCoefficient*>(Q);
+   FunctionCoefficient *function_coeff = dynamic_cast<FunctionCoefficient*>(Q);
+   // TODO: other types of coefficients ...
+   if (dim==1) { mfem_error("Not supported yet... stay tuned!"); }
+   if (dim==2)
+   {
+      double constant = 0.0;
+      double (*function)(const DeviceVector3&) = NULL;
+      if (const_coeff)
+      {
+         constant = const_coeff->constant;
+      }
+      else if (function_coeff)
+      {
+         function = function_coeff->GetFunction();
+      }
+      else
+      {
+         MFEM_ABORT("Coefficient type not supported");
+      }
+      const int NE = ne;
+      const int NQ = nq;
+      const int dims = el.GetDim();
+      const DeviceVector w(maps->W.GetData(), NQ);
+      const DeviceTensor<3> x(geo->X.GetData(), 3,NQ,NE);
+      const DeviceTensor<4> J(geo->J.GetData(), 2,2,NQ,NE);
+      DeviceMatrix v(vec.GetData(), NQ, NE);
+      MFEM_FORALL(e, NE,
+      {
+         for (int q = 0; q < NQ; ++q)
+         {
+            const double J11 = J(0,0,q,e);
+            const double J12 = J(1,0,q,e);
+            const double J21 = J(0,1,q,e);
+            const double J22 = J(1,1,q,e);
+            const double detJ = (J11*J22)-(J21*J12);
+            const int offset = dims*NQ*e+q;
+            const double coeff =
+            const_coeff ? constant:
+            function_coeff ?
+            function(DeviceVector3(x[offset], x[offset+1], x[offset+2])):
+            0.0;
+            v(q,e) =  w[q] * coeff * detJ;
+         }
+      });
+   }
+   if (dim==3)
+   {
+      double constant = 0.0;
+      double (*function)(const DeviceVector3&) = NULL;
+      if (const_coeff)
+      {
+         constant = const_coeff->constant;
+      }
+      else if (function_coeff)
+      {
+         function = function_coeff->GetFunction();
+      }
+      else
+      {
+         MFEM_ABORT("Coefficient type not supported");
+      }
+      const int NE = ne;
+      const int NQ = nq;
+      const int dims = el.GetDim();
+      const DeviceVector W(maps->W.GetData(), NQ);
+      const DeviceTensor<3> x(geo->X.GetData(), 3,NQ,NE);
+      const DeviceTensor<4> J(geo->J.GetData(), 3,3,NQ,NE);
+      DeviceMatrix v(vec.GetData(), NQ,NE);
+      MFEM_FORALL(e, NE,
+      {
+         for (int q = 0; q < NQ; ++q)
+         {
+            const double J11 = J(0,0,q,e),J12 = J(1,0,q,e),J13 = J(2,0,q,e);
+            const double J21 = J(0,1,q,e),J22 = J(1,1,q,e),J23 = J(2,1,q,e);
+            const double J31 = J(0,2,q,e),J32 = J(1,2,q,e),J33 = J(2,2,q,e);
+            const double detJ =
+            ((J11 * J22 * J33) + (J12 * J23 * J31) + (J13 * J21 * J32) -
+            (J13 * J22 * J31) - (J12 * J21 * J33) - (J11 * J23 * J32));
+            const int offset = dims*NQ*e+q;
+            const double coeff =
+            const_coeff ? constant:
+            function_coeff ?
+            function(DeviceVector3(x[offset], x[offset+1], x[offset+2])):
+            0.0;
+            v(q,e) = W(q) * coeff * detJ;
+         }
+      });
+   }
+   //delete geo;
+}
+
+// *****************************************************************************
+void PAMassIntegrator::MultAdd(Vector &x, Vector &y)
+{
+   kernels::fem::MassMultAssembled(dim, dofs1D, quad1D, ne,
+                                   maps->B, maps->Bt,
+                                   vec, x, y);
+}
+
+// ***************************************************************************
+// * DofToQuad
+// ***************************************************************************
+static std::map<std::string, DofToQuad* > AllDofQuadMaps;
+
+// *****************************************************************************
+DofToQuad* DofToQuad::Get(const FiniteElementSpace& fes,
+                          const IntegrationRule& ir,
+                          const bool transpose)
+{
+   return Get(*fes.GetFE(0), *fes.GetFE(0), ir, transpose);
+}
+
+DofToQuad* DofToQuad::Get(const FiniteElementSpace& trialFES,
+                          const FiniteElementSpace& testFES,
+                          const IntegrationRule& ir,
+                          const bool transpose)
+{
+   return Get(*trialFES.GetFE(0), *testFES.GetFE(0), ir, transpose);
+}
+
+DofToQuad* DofToQuad::Get(const FiniteElement& trialFE,
+                          const FiniteElement& testFE,
+                          const IntegrationRule& ir,
+                          const bool transpose)
+{
+   return GetTensorMaps(trialFE, testFE, ir, transpose);
+}
+
+// ***************************************************************************
+DofToQuad* DofToQuad::GetTensorMaps(const FiniteElement& trialFE,
+                                    const FiniteElement& testFE,
+                                    const IntegrationRule& ir,
+                                    const bool transpose)
+{
+   const TensorBasisElement& trialTFE =
+      dynamic_cast<const TensorBasisElement&>(trialFE);
+   const TensorBasisElement& testTFE =
+      dynamic_cast<const TensorBasisElement&>(testFE);
+   std::stringstream ss;
+   ss << "TensorMap:"
+      << " O1:"  << trialFE.GetOrder()
+      << " O2:"  << testFE.GetOrder()
+      << " BT1:" << trialTFE.GetBasisType()
+      << " BT2:" << testTFE.GetBasisType()
+      << " Q:"   << ir.GetNPoints();
+   std::string hash = ss.str();
+   // If we've already made the dof-quad maps, reuse them
+   if (AllDofQuadMaps.find(hash)!=AllDofQuadMaps.end())
+   {
+      return AllDofQuadMaps[hash];
+   }
+   // Otherwise, build them
+   DofToQuad *maps = new DofToQuad();
+   AllDofQuadMaps[hash]=maps;
+   maps->hash = hash;
+   const DofToQuad* trialMaps = GetD2QTensorMaps(trialFE, ir);
+   const DofToQuad* testMaps  = GetD2QTensorMaps(testFE, ir, true);
+   maps->B = trialMaps->B;
+   maps->G = trialMaps->G;
+   maps->Bt = testMaps->B;
+   maps->Gt = testMaps->G;
+   maps->W = testMaps->W;
+   delete trialMaps;
+   delete testMaps;
+   return maps;
+}
+
+// ***************************************************************************
+DofToQuad* DofToQuad::GetD2QTensorMaps(const FiniteElement& fe,
+                                       const IntegrationRule& ir,
+                                       const bool transpose)
+{
+   const IntegrationRule& ir1D = IntRules.Get(Geometry::SEGMENT,ir.GetOrder());
+
+   const int dims = fe.GetDim();
+   const int order = fe.GetOrder();
+   const int numDofs = order + 1;
+   const int numQuad1D = ir1D.GetNPoints();
+   const int numQuad2D = numQuad1D * numQuad1D;
+   const int numQuad3D = numQuad2D * numQuad1D;
+   const int numQuad =
+      (dims == 1) ? numQuad1D :
+      (dims == 2) ? numQuad2D :
+      (dims == 3) ? numQuad3D : 0;
+   assert(numQuad > 0);
+   std::stringstream ss;
+   ss << "D2QTensorMap:"
+      << " dims:" << dims
+      << " order:" << order
+      << " numDofs:" << numDofs
+      << " numQuad1D:" << numQuad1D
+      << " transpose:"  << (transpose?"true":"false");
+   std::string hash = ss.str();
+   if (AllDofQuadMaps.find(hash)!=AllDofQuadMaps.end())
+   {
+      return AllDofQuadMaps[hash];
+   }
+   DofToQuad *maps = new DofToQuad();
+   AllDofQuadMaps[hash]=maps;
+   maps->hash = hash;
+   maps->B.SetSize(numQuad1D*numDofs);
+   maps->G.SetSize(numQuad1D*numDofs);
+   const int dim0 = (!transpose)?1:numDofs;
+   const int dim1 = (!transpose)?numQuad1D:1;
+   if (transpose) // Initialize quad weights only for transpose
+   {
+      maps->W.SetSize(numQuad);
+   }
+   mfem::Vector d2q(numDofs);
+   mfem::Vector d2qD(numDofs);
+   mfem::Array<double> W1d(numQuad1D);
+   mfem::Array<double> B1d(numQuad1D*numDofs);
+   mfem::Array<double> G1d(numQuad1D*numDofs);
+   const TensorBasisElement& tbe = dynamic_cast<const TensorBasisElement&>(fe);
+   const Poly_1D::Basis& basis = tbe.GetBasis1D();
+   for (int q = 0; q < numQuad1D; ++q)
+   {
+      const IntegrationPoint& ip = ir1D.IntPoint(q);
+      if (transpose)
+      {
+         W1d[q] = ip.weight;
+      }
+      basis.Eval(ip.x, d2q, d2qD);
+      for (int d = 0; d < numDofs; ++d)
+      {
+         const double w = d2q[d];
+         const double wD = d2qD[d];
+         const int idx = dim0*q + dim1*d;
+         B1d[idx] = w;
+         G1d[idx] = wD;
+      }
+   }
+   if (transpose)
+   {
+      mfem::Array<double> W(numQuad);
+      for (int q = 0; q < numQuad; ++q)
+      {
+         const int qx = q % numQuad1D;
+         const int qz = q / numQuad2D;
+         const int qy = (q - qz*numQuad2D) / numQuad1D;
+         double w = W1d[qx];
+         if (dims > 1) { w *= W1d[qy]; }
+         if (dims > 2) { w *= W1d[qz]; }
+         W[q] = w;
+      }
+      maps->W = W;
+   }
+   kernels::vector::Assign(numQuad1D*numDofs, B1d, maps->B);
+   kernels::vector::Assign(numQuad1D*numDofs, G1d, maps->G);
+   return maps;
+}
+
+// ***************************************************************************
+DofToQuad* DofToQuad::GetSimplexMaps(const FiniteElement& fe,
+                                     const IntegrationRule& ir,
+                                     const bool transpose)
+{
+   return GetSimplexMaps(fe, fe, ir, transpose);
+}
+
+// *****************************************************************************
+DofToQuad* DofToQuad::GetSimplexMaps(const FiniteElement& trialFE,
+                                     const FiniteElement& testFE,
+                                     const IntegrationRule& ir,
+                                     const bool transpose)
+{
+   std::stringstream ss;
+   ss << "SimplexMap:"
+      << " O1:" << trialFE.GetOrder()
+      << " O2:" << testFE.GetOrder()
+      << " Q:"  << ir.GetNPoints();
+   std::string hash = ss.str();
+   // If we've already made the dof-quad maps, reuse them
+   if (AllDofQuadMaps.find(hash)!=AllDofQuadMaps.end())
+   {
+      return AllDofQuadMaps[hash];
+   }
+   DofToQuad *maps = new DofToQuad();
+   AllDofQuadMaps[hash]=maps;
+   maps->hash = hash;
+   const DofToQuad* trialMaps = GetD2QSimplexMaps(trialFE, ir);
+   const DofToQuad* testMaps  = GetD2QSimplexMaps(testFE, ir, true);
+   maps->B = trialMaps->B;
+   maps->G = trialMaps->G;
+   maps->Bt = testMaps->B;
+   maps->Gt = testMaps->G;
+   maps->W = testMaps->W;
+   delete trialMaps;
+   delete testMaps;
+   return maps;
+}
+
+// ***************************************************************************
+DofToQuad* DofToQuad::GetD2QSimplexMaps(const FiniteElement& fe,
+                                        const IntegrationRule& ir,
+                                        const bool transpose)
+{
+   const int dims = fe.GetDim();
+   const int numDofs = fe.GetDof();
+   const int numQuad = ir.GetNPoints();
+   std::stringstream ss ;
+   ss << "D2QSimplexMap:"
+      << " Dim:" << dims
+      << " numDofs:" << numDofs
+      << " numQuad:" << numQuad
+      << " transpose:" << (transpose?"true":"false");
+   std::string hash = ss.str();
+   if (AllDofQuadMaps.find(hash)!=AllDofQuadMaps.end())
+   {
+      return AllDofQuadMaps[hash];
+   }
+   DofToQuad* maps = new DofToQuad();
+   AllDofQuadMaps[hash]=maps;
+   maps->hash = hash;
+   maps->B.SetSize(numQuad*numDofs);
+   maps->G.SetSize(dims*numQuad*numDofs);
+   const int dim0 = (!transpose)?1:numDofs;
+   const int dim1 = (!transpose)?numQuad:1;
+   const int dim0D = (!transpose)?1:numQuad;
+   const int dim1D = (!transpose)?dims:1;
+   const int dim2D = dims*numQuad;
+   if (transpose) // Initialize quad weights only for transpose
+   {
+      maps->W.SetSize(numQuad);
+   }
+   mfem::Vector d2q(numDofs);
+   mfem::DenseMatrix d2qD(numDofs, dims);
+   mfem::Array<double> W(numQuad);
+   mfem::Array<double> B(numQuad*numDofs);
+   mfem::Array<double> G(dims*numQuad*numDofs);
+   for (int q = 0; q < numQuad; ++q)
+   {
+      const IntegrationPoint& ip = ir.IntPoint(q);
+      if (transpose)
+      {
+         W[q] = ip.weight;
+      }
+      fe.CalcShape(ip, d2q);
+      fe.CalcDShape(ip, d2qD);
+      for (int d = 0; d < numDofs; ++d)
+      {
+         const double w = d2q[d];
+         const int idx = dim0*q + dim1*d;
+         B[idx] = w;
+         for (int dim = 0; dim < dims; ++dim)
+         {
+            const double wD = d2qD(d, dim);
+            const int idxD = dim0D*dim + dim1D*q + dim2D*d;
+            G[idxD] = wD;
+         }
+      }
+   }
+   if (transpose)
+   {
+      kernels::vector::Assign(numQuad, W, maps->W);
+   }
+   kernels::vector::Assign(numQuad*numDofs, B, maps->B);
+   kernels::vector::Assign(dims*numQuad*numDofs, G, maps->G);
+   return maps;
+}
+
+// *****************************************************************************
+static long sequence = -1;
+static GeometryExtension *geom = NULL;
+
+// *****************************************************************************
+static void GeomFill(const int vdim,
+                     const size_t NE, const size_t ND, const size_t NX,
+                     const int* elementMap, int* eMap,
+                     const double *_X, double *meshNodes)
+{
+   const DeviceArray d_elementMap(elementMap, ND*NE);
+   DeviceArray d_eMap(eMap, ND*NE);
+   const DeviceVector X(_X, NX);
+   DeviceVector d_meshNodes(meshNodes, vdim*ND*NE);
+   MFEM_FORALL(e, NE,
+   {
+      for (size_t d = 0; d < ND; ++d)
+      {
+         const int lid = d+ND*e;
+         const int gid = d_elementMap[lid];
+         d_eMap[lid] = gid;
+         for (int v = 0; v < vdim; ++v)
+         {
+            const int moffset = v+vdim*lid;
+            const int xoffset = v+vdim*gid;
+            d_meshNodes[moffset] = X[xoffset];
+         }
+      }
+   });
+}
+
+// *****************************************************************************
+static void NodeCopyByVDim(const int elements,
+                           const int numDofs,
+                           const int ndofs,
+                           const int dims,
+                           const int* eMap,
+                           const double* Sx,
+                           double* nodes)
+{
+   MFEM_FORALL(e,elements,
+   {
+      for (int dof = 0; dof < numDofs; ++dof)
+      {
+         const int lid = dof+numDofs*e;
+         const int gid = eMap[lid];
+         for (int v = 0; v < dims; ++v)
+         {
+            const int moffset = v+dims*lid;
+            const int voffset = gid+v*ndofs;
+            nodes[moffset] = Sx[voffset];
+         }
+      }
+   });
+}
+
+
+// *****************************************************************************
+GeometryExtension* GeometryExtension::Get(const FiniteElementSpace& fes,
+                                          const IntegrationRule& ir,
+                                          const Vector& Sx)
+{
+   const Mesh *mesh = fes.GetMesh();
+   const GridFunction *nodes = mesh->GetNodes();
+   const FiniteElementSpace *fespace = nodes->FESpace();
+   const FiniteElement *fe = fespace->GetFE(0);
+   const int dims     = fe->GetDim();
+   const int numDofs  = fe->GetDof();
+   const int numQuad  = ir.GetNPoints();
+   const int elements = fespace->GetNE();
+   const int ndofs    = fespace->GetNDofs();
+   const DofToQuad* maps = DofToQuad::GetSimplexMaps(*fe, ir);
+   NodeCopyByVDim(elements,numDofs,ndofs,dims,geom->eMap,Sx,geom->nodes);
+   kernels::fem::Geom(dims, numDofs, numQuad, elements,
+                      maps->G, geom->nodes,
+                      geom->X, geom->J, geom->invJ, geom->detJ);
+   return geom;
+}
+
+// *****************************************************************************
+GeometryExtension* GeometryExtension::Get(const FiniteElementSpace& fes,
+                                          const IntegrationRule& ir)
+{
+   Mesh *mesh = fes.GetMesh();
+   const bool geom_to_allocate = sequence < fes.GetSequence();
+   sequence = fes.GetSequence();
+   if (geom_to_allocate) { geom = new GeometryExtension(); }
+   mesh->EnsureNodes();
+   const GridFunction *nodes = mesh->GetNodes();
+   const mfem::FiniteElementSpace *fespace = nodes->FESpace();
+   const mfem::FiniteElement *fe = fespace->GetFE(0);
+   const int dims     = fe->GetDim();
+   const int elements = fespace->GetNE();
+   const int numDofs  = fe->GetDof();
+   const int numQuad  = ir.GetNPoints();
+   const bool orderedByNODES = (fespace->GetOrdering() == Ordering::byNODES);
+   if (orderedByNODES) { ReorderByVDim(nodes); }
+   const int asize = dims*numDofs*elements;
+   mfem::Array<double> meshNodes(asize);
+   const Table& e2dTable = fespace->GetElementToDofTable();
+   const int* elementMap = e2dTable.GetJ();
+   mfem::Array<int> eMap(numDofs*elements);
+   GeomFill(dims,
+            elements,
+            numDofs,
+            nodes->Size(),
+            elementMap,
+            eMap,
+            nodes->GetData(),
+            meshNodes);
+   if (geom_to_allocate)
+   {
+      geom->nodes.SetSize(dims*numDofs*elements);
+      geom->eMap.SetSize(numDofs*elements);
+   }
+   geom->nodes = meshNodes;
+   geom->eMap = eMap;
+   // Reorder the original gf back
+   if (orderedByNODES) { ReorderByNodes(nodes); }
+   if (geom_to_allocate)
+   {
+      geom->X.SetSize(dims*numQuad*elements);
+      geom->J.SetSize(dims*dims*numQuad*elements);
+      geom->invJ.SetSize(dims*dims*numQuad*elements);
+      geom->detJ.SetSize(numQuad*elements);
+   }
+   const DofToQuad* maps = DofToQuad::GetSimplexMaps(*fe, ir);
+   kernels::fem::Geom(dims, numDofs, numQuad, elements,
+                      maps->G, geom->nodes,
+                      geom->X, geom->J, geom->invJ, geom->detJ);
+   return geom;
+}
+
+// ***************************************************************************
+void GeometryExtension::ReorderByVDim(const GridFunction *nodes)
+{
+   const mfem::FiniteElementSpace *fes = nodes->FESpace();
+   const int size = nodes->Size();
+   const int vdim = fes->GetVDim();
+   const int ndofs = fes->GetNDofs();
+   double *data = nodes->GetData();
+   double *temp = new double[size];
+   int k=0;
+   for (int d = 0; d < ndofs; d++)
+      for (int v = 0; v < vdim; v++)
+      {
+         temp[k++] = data[d+v*ndofs];
+      }
+   for (int i = 0; i < size; i++)
+   {
+      data[i] = temp[i];
+   }
+   delete [] temp;
+}
+
+// ***************************************************************************
+void GeometryExtension::ReorderByNodes(const GridFunction *nodes)
+{
+   const mfem::FiniteElementSpace *fes = nodes->FESpace();
+   const int size = nodes->Size();
+   const int vdim = fes->GetVDim();
+   const int ndofs = fes->GetNDofs();
+   double *data = nodes->GetData();
+   double *temp = new double[size];
+   int k = 0;
+   for (int j = 0; j < ndofs; j++)
+      for (int i = 0; i < vdim; i++)
+      {
+         temp[j+i*ndofs] = data[k++];
+      }
+   for (int i = 0; i < size; i++)
+   {
+      data[i] = temp[i];
+   }
+   delete [] temp;
+}
+
+} // namespace mfem
