+++ fem/bilininteg_ext.hpp
@@ -9,14 +9,32 @@
 // terms of the GNU Lesser General Public License (as published by the Free
 // Software Foundation) version 2.1 dated February 1999.
 
-#ifndef MFEM_BILIN_DQM_HPP
-#define MFEM_BILIN_DQM_HPP
-
-#include "../config/config.hpp"
-#include "../general/array.hpp"
+#ifndef MFEM_BILININTEG_EXT
+#define MFEM_BILININTEG_EXT
 
 namespace mfem
 {
+
+// ***************************************************************************
+// * GeometryExtension
+// ***************************************************************************
+class GeometryExtension
+{
+public:
+   Array<int> eMap;
+   Array<double> nodes;
+   Array<double> X, J, invJ, detJ;
+   static const int Jacobian    = (1 << 0);
+   static const int JacobianInv = (1 << 1);
+   static const int JacobianDet = (1 << 2);
+   static GeometryExtension* Get(const FiniteElementSpace&,
+                                 const IntegrationRule&);
+   static GeometryExtension* Get(const FiniteElementSpace&,
+                                 const IntegrationRule&,
+                                 const Vector&);
+   static void ReorderByVDim(const GridFunction*);
+   static void ReorderByNodes(const GridFunction*);
+};
 
 // ***************************************************************************
 // * DofToQuad
@@ -59,6 +77,6 @@ public:
                                        const bool = false);
 };
 
-} // namespace mfem
+}
 
-#endif // MFEM_BACKENDS_KERNELS_BILIN_DQM_HPP
+#endif
