--- fem/doftoquad.cpp
@@ -1,286 +1,0 @@
-// Copyright (c) 2010, Lawrence Livermore National Security, LLC. Produced at
-// the Lawrence Livermore National Laboratory. LLNL-CODE-443211. All Rights
-// reserved. See file COPYRIGHT for details.
-//
-// This file is part of the MFEM library. For more information and source code
-// availability see http://mfem.org.
-//
-// MFEM is free software; you can redistribute it and/or modify it under the
-// terms of the GNU Lesser General Public License (as published by the Free
-// Software Foundation) version 2.1 dated February 1999.
-
-#include "../config/config.hpp"
-
-#include "../general/okina.hpp"
-#include "../linalg/kernels/vector.hpp"
-#include "fem.hpp"
-
-#include "doftoquad.hpp"
-#include <map>
-
-namespace mfem
-{
-
-// ***************************************************************************
-// * DofToQuad
-// ***************************************************************************
-static std::map<std::string, DofToQuad* > AllDofQuadMaps;
-
-// *****************************************************************************
-DofToQuad* DofToQuad::Get(const FiniteElementSpace& fes,
-                          const IntegrationRule& ir,
-                          const bool transpose)
-{
-   return Get(*fes.GetFE(0), *fes.GetFE(0), ir, transpose);
-}
-
-DofToQuad* DofToQuad::Get(const FiniteElementSpace& trialFES,
-                          const FiniteElementSpace& testFES,
-                          const IntegrationRule& ir,
-                          const bool transpose)
-{
-   return Get(*trialFES.GetFE(0), *testFES.GetFE(0), ir, transpose);
-}
-
-DofToQuad* DofToQuad::Get(const FiniteElement& trialFE,
-                          const FiniteElement& testFE,
-                          const IntegrationRule& ir,
-                          const bool transpose)
-{
-   return GetTensorMaps(trialFE, testFE, ir, transpose);
-}
-
-// ***************************************************************************
-DofToQuad* DofToQuad::GetTensorMaps(const FiniteElement& trialFE,
-                                    const FiniteElement& testFE,
-                                    const IntegrationRule& ir,
-                                    const bool transpose)
-{
-   const TensorBasisElement& trialTFE =
-      dynamic_cast<const TensorBasisElement&>(trialFE);
-   const TensorBasisElement& testTFE =
-      dynamic_cast<const TensorBasisElement&>(testFE);
-   std::stringstream ss;
-   ss << "TensorMap:"
-      << " O1:"  << trialFE.GetOrder()
-      << " O2:"  << testFE.GetOrder()
-      << " BT1:" << trialTFE.GetBasisType()
-      << " BT2:" << testTFE.GetBasisType()
-      << " Q:"   << ir.GetNPoints();
-   std::string hash = ss.str();
-   // If we've already made the dof-quad maps, reuse them
-   if (AllDofQuadMaps.find(hash)!=AllDofQuadMaps.end())
-   {
-      return AllDofQuadMaps[hash];
-   }
-   // Otherwise, build them
-   DofToQuad *maps = new DofToQuad();
-   AllDofQuadMaps[hash]=maps;
-   maps->hash = hash;
-   const DofToQuad* trialMaps = GetD2QTensorMaps(trialFE, ir);
-   const DofToQuad* testMaps  = GetD2QTensorMaps(testFE, ir, true);
-   maps->B = trialMaps->B;
-   maps->G = trialMaps->G;
-   maps->Bt = testMaps->B;
-   maps->Gt = testMaps->G;
-   maps->W = testMaps->W;
-   delete trialMaps;
-   delete testMaps;
-   return maps;
-}
-
-// ***************************************************************************
-DofToQuad* DofToQuad::GetD2QTensorMaps(const FiniteElement& fe,
-                                       const IntegrationRule& ir,
-                                       const bool transpose)
-{
-   const IntegrationRule& ir1D = IntRules.Get(Geometry::SEGMENT,ir.GetOrder());
-
-   const int dims = fe.GetDim();
-   const int order = fe.GetOrder();
-   const int numDofs = order + 1;
-   const int numQuad1D = ir1D.GetNPoints();
-   const int numQuad2D = numQuad1D * numQuad1D;
-   const int numQuad3D = numQuad2D * numQuad1D;
-   const int numQuad =
-      (dims == 1) ? numQuad1D :
-      (dims == 2) ? numQuad2D :
-      (dims == 3) ? numQuad3D : 0;
-   assert(numQuad > 0);
-   std::stringstream ss;
-   ss << "D2QTensorMap:"
-      << " dims:" << dims
-      << " order:" << order
-      << " numDofs:" << numDofs
-      << " numQuad1D:" << numQuad1D
-      << " transpose:"  << (transpose?"true":"false");
-   std::string hash = ss.str();
-   if (AllDofQuadMaps.find(hash)!=AllDofQuadMaps.end())
-   {
-      return AllDofQuadMaps[hash];
-   }
-   DofToQuad *maps = new DofToQuad();
-   AllDofQuadMaps[hash]=maps;
-   maps->hash = hash;
-   maps->B.SetSize(numQuad1D*numDofs);
-   maps->G.SetSize(numQuad1D*numDofs);
-   const int dim0 = (!transpose)?1:numDofs;
-   const int dim1 = (!transpose)?numQuad1D:1;
-   if (transpose) // Initialize quad weights only for transpose
-   {
-      maps->W.SetSize(numQuad);
-   }
-   mfem::Vector d2q(numDofs);
-   mfem::Vector d2qD(numDofs);
-   mfem::Array<double> W1d(numQuad1D);
-   mfem::Array<double> B1d(numQuad1D*numDofs);
-   mfem::Array<double> G1d(numQuad1D*numDofs);
-   const TensorBasisElement& tbe = dynamic_cast<const TensorBasisElement&>(fe);
-   const Poly_1D::Basis& basis = tbe.GetBasis1D();
-   for (int q = 0; q < numQuad1D; ++q)
-   {
-      const IntegrationPoint& ip = ir1D.IntPoint(q);
-      if (transpose)
-      {
-         W1d[q] = ip.weight;
-      }
-      basis.Eval(ip.x, d2q, d2qD);
-      for (int d = 0; d < numDofs; ++d)
-      {
-         const double w = d2q[d];
-         const double wD = d2qD[d];
-         const int idx = dim0*q + dim1*d;
-         B1d[idx] = w;
-         G1d[idx] = wD;
-      }
-   }
-   if (transpose)
-   {
-      mfem::Array<double> W(numQuad);
-      for (int q = 0; q < numQuad; ++q)
-      {
-         const int qx = q % numQuad1D;
-         const int qz = q / numQuad2D;
-         const int qy = (q - qz*numQuad2D) / numQuad1D;
-         double w = W1d[qx];
-         if (dims > 1) { w *= W1d[qy]; }
-         if (dims > 2) { w *= W1d[qz]; }
-         W[q] = w;
-      }
-      maps->W = W;
-   }
-   kernels::vector::Assign(numQuad1D*numDofs, B1d, maps->B);
-   kernels::vector::Assign(numQuad1D*numDofs, G1d, maps->G);
-   return maps;
-}
-
-// ***************************************************************************
-DofToQuad* DofToQuad::GetSimplexMaps(const FiniteElement& fe,
-                                     const IntegrationRule& ir,
-                                     const bool transpose)
-{
-   return GetSimplexMaps(fe, fe, ir, transpose);
-}
-
-// *****************************************************************************
-DofToQuad* DofToQuad::GetSimplexMaps(const FiniteElement& trialFE,
-                                     const FiniteElement& testFE,
-                                     const IntegrationRule& ir,
-                                     const bool transpose)
-{
-   std::stringstream ss;
-   ss << "SimplexMap:"
-      << " O1:" << trialFE.GetOrder()
-      << " O2:" << testFE.GetOrder()
-      << " Q:"  << ir.GetNPoints();
-   std::string hash = ss.str();
-   // If we've already made the dof-quad maps, reuse them
-   if (AllDofQuadMaps.find(hash)!=AllDofQuadMaps.end())
-   {
-      return AllDofQuadMaps[hash];
-   }
-   DofToQuad *maps = new DofToQuad();
-   AllDofQuadMaps[hash]=maps;
-   maps->hash = hash;
-   const DofToQuad* trialMaps = GetD2QSimplexMaps(trialFE, ir);
-   const DofToQuad* testMaps  = GetD2QSimplexMaps(testFE, ir, true);
-   maps->B = trialMaps->B;
-   maps->G = trialMaps->G;
-   maps->Bt = testMaps->B;
-   maps->Gt = testMaps->G;
-   maps->W = testMaps->W;
-   delete trialMaps;
-   delete testMaps;
-   return maps;
-}
-
-// ***************************************************************************
-DofToQuad* DofToQuad::GetD2QSimplexMaps(const FiniteElement& fe,
-                                        const IntegrationRule& ir,
-                                        const bool transpose)
-{
-   const int dims = fe.GetDim();
-   const int numDofs = fe.GetDof();
-   const int numQuad = ir.GetNPoints();
-   std::stringstream ss ;
-   ss << "D2QSimplexMap:"
-      << " Dim:" << dims
-      << " numDofs:" << numDofs
-      << " numQuad:" << numQuad
-      << " transpose:" << (transpose?"true":"false");
-   std::string hash = ss.str();
-   if (AllDofQuadMaps.find(hash)!=AllDofQuadMaps.end())
-   {
-      return AllDofQuadMaps[hash];
-   }
-   DofToQuad* maps = new DofToQuad();
-   AllDofQuadMaps[hash]=maps;
-   maps->hash = hash;
-   maps->B.SetSize(numQuad*numDofs);
-   maps->G.SetSize(dims*numQuad*numDofs);
-   const int dim0 = (!transpose)?1:numDofs;
-   const int dim1 = (!transpose)?numQuad:1;
-   const int dim0D = (!transpose)?1:numQuad;
-   const int dim1D = (!transpose)?dims:1;
-   const int dim2D = dims*numQuad;
-   if (transpose) // Initialize quad weights only for transpose
-   {
-      maps->W.SetSize(numQuad);
-   }
-   mfem::Vector d2q(numDofs);
-   mfem::DenseMatrix d2qD(numDofs, dims);
-   mfem::Array<double> W(numQuad);
-   mfem::Array<double> B(numQuad*numDofs);
-   mfem::Array<double> G(dims*numQuad*numDofs);
-   for (int q = 0; q < numQuad; ++q)
-   {
-      const IntegrationPoint& ip = ir.IntPoint(q);
-      if (transpose)
-      {
-         W[q] = ip.weight;
-      }
-      fe.CalcShape(ip, d2q);
-      fe.CalcDShape(ip, d2qD);
-      for (int d = 0; d < numDofs; ++d)
-      {
-         const double w = d2q[d];
-         const int idx = dim0*q + dim1*d;
-         B[idx] = w;
-         for (int dim = 0; dim < dims; ++dim)
-         {
-            const double wD = d2qD(d, dim);
-            const int idxD = dim0D*dim + dim1D*q + dim2D*d;
-            G[idxD] = wD;
-         }
-      }
-   }
-   if (transpose)
-   {
-      kernels::vector::Assign(numQuad, W, maps->W);
-   }
-   kernels::vector::Assign(numQuad*numDofs, B, maps->B);
-   kernels::vector::Assign(dims*numQuad*numDofs, G, maps->G);
-   return maps;
-}
-
-} // namespace mfem
